Struct rust_fc::io::FC

pub struct FC {
    // some fields omitted
}

Flight Computer IO.

Internally holds state for this implementation of the flight computer. This includes time (nanosecond counter from startup), a socket for listening for incoming data, a socket for sending data over the telemetry link, a log file, a running count of telemetry messages sent and a buffer for partly built telemetry messages.

To initialize use the Default trait:

Example

use rust_fc::io;

let mut flight_computer: io::FC = Default::default();

Methods

impl FC

fn listen(&self) -> Option<(u32, u16, Duration, [u8; P_LIMIT - 4])>

Listen for messages from the network.

This makes a blocking read call on the fc_listen_socket, waiting for any message from the outside world. Once received, it will deal with the sequence numbers in the header of the data and write the raw message to the passed in buffer.

Returns:

An Option containing a tuple of data from the socket. The tuple contains:

Sequence Number: Sequence number from the header of the data packet

Port: Which port the data was sent from

Time: The time that the message was received

Message: A message buffer that has raw bytes off the wire

Example

use rust_fc::io;

let mut flight_computer: io::FC = Default::default();

if let Some((seqn, recv_port, recv_time, message)) = flight_computer.listen() {
    // Do something here with received data
}

fn log_message(&mut self, message: &[u8], name: [u8; 4], time: Duration, message_size: usize) -> Result<(), Error>

Log a message to disk.

All data we care about can be encoded as a "message". The original code defined messages as packed structs in C. The reasoning was to be as space-efficient as reasonably possible given that we are both disk-size and bandwidth constrained.

This function takes a message (as an array of bytes) and writes it to disk.

Parameters:

message: Byte array containing packed message

name: Byte array of the name for this message

time: Time of message

message_size: How many bytes to copy from the message array

Returns:

A Result with any errors. But we hope to never deal with a failure here (Greater care was taken in the original flight computer to not crash because of disk errors).

fn telemetry(&mut self, message: &[u8], name: [u8; 4], time: Duration, message_size: usize)

Send a message to the ground.

All data we care about can be encoded as a "message". The original code defined messages as packed structs in C. The reasoning was to be as space-efficient as reasonably possible given that we are both disk-size and bandwidth constrained.

This function takes a message (as an array of bytes) and queues it to be send out over the network once will fill the maximum size of a UDP packet.

Parameters

message: Byte array containing packed message

name: Byte array of the name for this message

time: Time of message

message_size: How many bytes to copy from the message array