Struct rust_fc::io::SequenceError

pub struct SequenceError {
    pub port: u16,
    pub expected: u32,
    pub received: u32,
}

A sequence error message.

When we miss a packet or get an out of order packet we should log that for future analysis. This stores a error and builds a log-able message.

Example

use rust_fc::io;

let seqerror = io::SequenceError {
    port: 35020,
    expected: 12345,
    received: 12349,
};

// Now you can log or send the message somewhere

Fields

port

Which port the packet error is from

expected

Expected packet sequence number

received

Actual received packet sequence number

Methods

impl SequenceError

fn as_message(&self) -> [u8; SIZE_OF_SEQE]

Return a copy of this struct as a byte array.

The PSAS file and message type is always a byte array with big-endian representation of fields in a struct.