Struct rust_fc::state::State

pub struct State {
    pub time: u64,
    pub acc_up: f64,
    pub vel_up: f64,
    pub altitude: f64,
    pub roll_rate: f64,
    pub roll_angle: f64,
}

A representation for the current state vector in MKS units.

Fields

time

Exact time (nanoseconds from boot) that state-vector is valid for.

acc_up

Vertical Acceleration [m/s²]

vel_up

Vertical velocity [m/s]

altitude

Altitude above sea level [m]

roll_rate

Roll rate [deg/s]

roll_angle

Roll angle deg

Methods

impl State

fn update_imu(&mut self, time: Duration, imu: ADIS)

Update the state based on new IMU data.

This is expected to be called every time new data is available from a sensor.

Parameters:

time: Time that new data is valid for

imu: New IMU data

fn as_message(&mut self) -> [u8; SIZE_OF_STATE]

Return a copy of this struct as a byte array.

The PSAS file and message type is always a byte array with big-endian representation of fields in a struct.